requirements.txt
appdirs==1.4.0
future==0.16.0
packaging==16.8
pyparsing==2.1.10
python-whois==0.6.5
requests==2.13.0
six==1.10.0
